split up rollup code
