+++ rollup/index.js
@@ -1,0 +1,8 @@
+Error.stackTraceLimit = Infinity;
+
+// TODO Is this necessary?
+import * as plugins from './plugins/index.js';
+export { plugins };
+
+export * from './bundleHelpers.js';
+export { insertAt } from './utils.js';
