+++ rollup/utils.js
@@ -1,0 +1,16 @@
+/**
+ * Helper functions to compensate for the fact that JS can't handle negative array indices very well
+ */
+
+export const getLastElement = array => {
+  return array[array.length - 1];
+};
+
+export const insertAt = (arr, index, ...insertees) => {
+  const newArr = [...arr];
+  // Add 1 to the array length so that the inserted element ends up in the right spot with respect to the length of the
+  // new array (which will be one element longer), rather than that of the current array
+  const destinationIndex = index >= 0 ? index : arr.length + 1 + index;
+  newArr.splice(destinationIndex, 0, ...insertees);
+  return newArr;
+};
